+++ entitlements.mas.inherit.plist
