+++ package.json
@@ -236,6 +236,7 @@
     "electron-is-dev": "^1.2.0",
     "electron-notarize": "^1.0.0",
     "electron-osx-sign": "^0.5.0",
+    "electron-routes": "^1.3.0",
     "electron-updater": "^4.3.5",
     "execa": "^5.0.0",
     "express": "^4.17.2",
