+++ signmasscript.sh
